dom/tests/mochitest/bugs/test_bug698061.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=698061
-->
<head>
  <title>Test for Bug 698061</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=698061">Mozilla Bug 698061</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 698061 **/

var d = document.createElement("div");
d.innerHTML = "<span>hello </span><span>world</span>";

var imported = document.importNode(d);
is(imported.childNodes.length, 0, "Should not have cloned child nodes with no deep arg!");
is(imported.textContent, "", "Should not have cloned text with no deep arg!");

imported = document.importNode(d, true);
is(imported.childNodes.length, 2, "Should have cloned child nodes!");
is(imported.textContent, "hello world", "Should have cloned text!");

imported = document.importNode(d, false);
is(imported.childNodes.length, 0, "Shouldn't have cloned child nodes!");
is(imported.textContent, "", "Shouldn't have cloned text!");

</script>
</pre>
</body>
</html>
